code/requirements.txt
python-steam-api
discord.py
python-dotenv
mcrcon